general restructuration
