+++ .env.dev.example
@@ -10,3 +10,4 @@ FIM_MANHA = HH:MM
 FIM_TARDE = HH:MM
 API_BASIC_USER=apiuser
 API_BASIC_PASS=apipassphrase
+AUTO_CREATE_DB = True
