+++ app/main.py
@@ -13,7 +13,9 @@ def create_app():
     with app.app_context():
         register_blueprints(app)
         from app.auxiliar import auxiliar_template
+        from app.auxiliar import error
         auxiliar_template.register_filters(app)
+        error.register_error_handler(app)
 
         if os.getenv("AUTO_CREATE_DB") == "True":
             db.create_all()
