+++ app/routes/database/aulas_ativas.py
@@ -9,4 +9,4 @@ bp = Blueprint('aulas_ativas', __name__, url_prefix="/admin")
 @admin_required
 def gerenciar_aulas_ativas():
     flash("Pagina em Desenvolvimento", "warning")
-    return redirect(url_for('error.under_dev_page'))
+    return redirect(url_for('default.under_dev_page'))
