+++ app/templates/auth/login.html
@@ -10,7 +10,7 @@
         <div class="panel-body">
             <form class="form"
                   role="form"
-                  action="{{ url_for('login') }}"
+                  action="{{ url_for('auth.login') }}"
                   method="post">
                 <!-- Campo de usuário -->
                 <div class="form-group">
