+++ app/templates/database/historicos.html
@@ -1,9 +1,9 @@
 {% extends "database/base_crude" %}
 {% block title %}
-    Historico
+    Historicos
 {% endblock title %}
 {% block content_head %}
-    {{ generate_head(url_for("gerenciar_Historico"), acao, disable=['inserir', 'editar',
+    {{ generate_head(url_for("historicos.gerenciar_Historicos"), acao, disable=['inserir', 'editar',
     'excluir']) }}
 {% endblock content_head %}
 {% block abertura %}
