+++ app/templates/database/permissoes.html
@@ -3,7 +3,7 @@
     Permissoes
 {% endblock title %}
 {% block content_head %}
-    {{ generate_head(url_for("gerenciar_permissoes"), acao) }}
+    {{ generate_head(url_for("permissoes.gerenciar_permissoes"), acao) }}
 {% endblock content_head %}
 {% block abertura %}
     <h3>
@@ -131,7 +131,7 @@
         <h3>Procurar Permissoes</h3>
         <form class="form-group"
               role="form"
-              action="{{ url_for('gerenciar_permissoes') }}"
+              action="{{ url_for('permissoes.gerenciar_permissoes') }}"
               method="post">
             <input type="hidden" name="acao" value="{{ acao }}">
             {{ gerar_select("id_usuario", "Selecione o usuario", users, "id_usuario", "nome_pessoa", "id_permissao_usuario", optional=True) 
@@ -246,7 +246,7 @@
     {% if users %}
         <form class="form-group"
               role="form"
-              action="{{ url_for('gerenciar_permissoes') }}"
+              action="{{ url_for('permissoes.gerenciar_permissoes') }}"
               method="post">
             <input type="hidden" name="acao" value="{{ acao }}">
             {{ gerar_select("id_usuario", "Selecione o usuario", users, "id_usuario", "nome_pessoa", "id_permissao_usuario") }}
@@ -286,7 +286,7 @@
     {% if bloco == 0 %}
         <form class="form-group"
               role="form"
-              action="{{ url_for('gerenciar_permissoes') }}"
+              action="{{ url_for('permissoes.gerenciar_permissoes') }}"
               method="post">
             <input type="hidden" name="acao" value="{{ acao }}">
             {{ gerar_select("id_usuario", "Selecione o usuario", permissoes, "id_permissao_usuario", "nome_pessoa", "id_usuario") }}
@@ -295,7 +295,7 @@
     {% elif bloco == 1 %}
         <form class="form-group"
               role="form"
-              action="{{ url_for('gerenciar_permissoes') }}"
+              action="{{ url_for('permissoes.gerenciar_permissoes') }}"
               method="post">
             <input type="hidden" name="acao" value="{{ acao }}">
             <div class="form-group">
@@ -353,7 +353,7 @@
     {% if bloco == 0 %}
         <form class="form-group"
               role="form"
-              action="{{ url_for('gerenciar_permissoes') }}"
+              action="{{ url_for('permissoes.gerenciar_permissoes') }}"
               method="post">
             <input type="hidden" name="acao" value="{{ acao }}">
             {{ gerar_select("id_usuario", "Selecione o usuario", permissoes, "id_permissao_usuario", "nome_pessoa", "id_usuario") }}
@@ -362,7 +362,7 @@
     {% elif bloco == 1 %}
         <form class="form-group"
               role="form"
-              action="{{ url_for('gerenciar_permissoes') }}"
+              action="{{ url_for('permissoes.gerenciar_permissoes') }}"
               method="post">
             <input type="hidden" name="acao" value="{{ acao }}">
             <div class="form-group">
