+++ app/templates/database/semestres.html
@@ -3,7 +3,7 @@
     Semestres
 {% endblock title %}
 {% block content_head %}
-    {{ generate_head(url_for("gerenciar_semestres") , acao) }}
+    {{ generate_head(url_for("semestres.gerenciar_semestres") , acao) }}
 {% endblock content_head %}
 {% block abertura %}
     <h3>
