+++ app/templates/database/usuarios.html
@@ -3,7 +3,7 @@
     Usuários
 {% endblock title %}
 {% block content_head %}
-    {{ generate_head(url_for("gerenciar_usuarios"), acao, disable = disable) }}
+    {{ generate_head(url_for("usuarios.gerenciar_usuarios"), acao, disable = disable) }}
 {% endblock content_head %}
 {% block abertura %}
     <h3>
@@ -127,7 +127,7 @@
         <h3>Procurar Usuário</h3>
         <form class="form-group"
               role="form"
-              action="{{ url_for('gerenciar_usuarios') }}"
+              action="{{ url_for('usuarios.gerenciar_usuarios') }}"
               method="post">
             <input type="hidden" name="acao" value="{{ acao }}">
             <div class="form-group">
@@ -248,7 +248,7 @@
     <h3>Inserir Usuários</h3>
     <form class="form-group"
           role="form"
-          action="{{ url_for('gerenciar_usuarios') }}"
+          action="{{ url_for('usuarios.gerenciar_usuarios') }}"
           method="post">
         <input type="hidden" name="acao" value="{{ acao }}">
         {{ gerar_select("pessoa", "Selecione a Pessoa", pessoas, "id_pessoa", "nome_pessoa") }}
@@ -303,7 +303,7 @@
         <h3>Editar Usuários</h3>
         <form class="form-group"
               role="form"
-              action="{{ url_for('gerenciar_usuarios') }}"
+              action="{{ url_for('usuarios.gerenciar_usuarios') }}"
               method="post">
             <input type="hidden" name="acao" value="{{ acao }}">
             {{ gerar_select("id_usuario", "Selecione o usuario", usuarios, "id_usuario", "nome_pessoa") }}
@@ -313,7 +313,7 @@
         <h3>Editar Usuário</h3>
         <form class="form-group"
               role="form"
-              action="{{ url_for('gerenciar_usuarios') }}"
+              action="{{ url_for('usuarios.gerenciar_usuarios') }}"
               method="post">
             <input type="hidden" name="acao" value="{{ acao }}">
             <div class="form-group">
@@ -383,7 +383,7 @@
         <h3>Excluir Usuários</h3>
         <form class="form-group"
               role="form"
-              action="{{ url_for('gerenciar_usuarios') }}"
+              action="{{ url_for('usuarios.gerenciar_usuarios') }}"
               method="post">
             <input type="hidden" name="acao" value="{{ acao }}">
             {{ gerar_select("id_usuario", "Selecione o usuario", usuarios, "id_usuario", "nome_pessoa") }}
@@ -393,7 +393,7 @@
         <h3>Excluir Usuário</h3>
         <form class="form-group"
               role="form"
-              action="{{ url_for('gerenciar_usuarios') }}"
+              action="{{ url_for('usuarios.gerenciar_usuarios') }}"
               method="post">
             <input type="hidden" name="acao" value="{{ acao }}">
             <div class="form-group">
