+++ app/templates/database/usuarios_especiais.html
@@ -3,7 +3,7 @@
     Usuarios Especiais
 {% endblock title %}
 {% block content_head %}
-    {{ generate_head(url_for("gerenciar_usuarios_especiais"), acao) }}
+    {{ generate_head(url_for("usuarios_especiais.gerenciar_usuarios_especiais"), acao) }}
 {% endblock content_head %}
 {% block abertura %}
     <h3>
@@ -94,7 +94,7 @@
         <h3>Procurar por Usuarios Especiais</h3>
         <form class="form-group"
               role="form"
-              action="{{ url_for('gerenciar_usuarios_especiais') }}"
+              action="{{ url_for('usuarios_especiais.gerenciar_usuarios_especiais') }}"
               method="post">
             <input type="hidden" name="acao" value="{{ acao }}">
             <div class="form-group">
@@ -180,7 +180,7 @@
     <h3>Inserir Usuarios Especiais</h3>
     <form class="form-group"
           role="form"
-          action="{{ url_for('gerenciar_usuarios_especiais') }}"
+          action="{{ url_for('usuarios_especiais.gerenciar_usuarios_especiais') }}"
           method="post">
         <input type="hidden" name="acao" value="{{ acao }}">
         <div class="form-group">
@@ -201,7 +201,7 @@
         <h3>Editar Usuarios Especiais</h3>
         <form class="form-group"
               role="form"
-              action="{{ url_for('gerenciar_usuarios_especiais') }}"
+              action="{{ url_for('usuarios_especiais.gerenciar_usuarios_especiais') }}"
               method="post">
             <input type="hidden" name="acao" value="{{ acao }}">
             {{ gerar_select("id_usuario_especial", "Selecione o usuario especial", usuarios_especiais, "id_usuario_especial", "nome_usuario_especial") }}
@@ -211,7 +211,7 @@
         <h3>Editar Usuario Especial</h3>
         <form class="form-group"
               role="form"
-              action="{{ url_for('gerenciar_usuarios_especiais') }}"
+              action="{{ url_for('usuarios_especiais.gerenciar_usuarios_especiais') }}"
               method="post">
             <input type="hidden" name="acao" value="{{ acao }}">
             <div class="form-group">
@@ -245,7 +245,7 @@
         <h3>Excluir Usuarios Especiais</h3>
         <form class="form-group"
               role="form"
-              action="{{ url_for('gerenciar_usuarios_especiais') }}"
+              action="{{ url_for('usuarios_especiais.gerenciar_usuarios_especiais') }}"
               method="post">
             <input type="hidden" name="acao" value="{{ acao }}">
             {{ gerar_select("id_usuario_especial", "Selecione o usuario especial", usuarios_especiais, "id_usuario_especial", "nome_usuario_especial") }}
@@ -255,7 +255,7 @@
         <h3>Excluir Usuario Especial</h3>
         <form class="form-group"
               role="form"
-              action="{{ url_for('gerenciar_usuarios_especiais') }}"
+              action="{{ url_for('usuarios_especiais.gerenciar_usuarios_especiais') }}"
               method="post">
             <input type="hidden" name="acao" value="{{ acao }}">
             <div class="form-group">
