+++ app/templates/http/401.html
@@ -9,6 +9,6 @@
     </div>
     <div class="btn-group">
         <a href="{{ url_for('default.home') }}" class="btn btn-info"><span class="glyphicon glyphicon-home"></span> Voltar à página de início</a>
-        <a href="{{ url_for('login') }}" class="btn btn-success"><span class="glyphicon glyphicon-log-in"></span> Ir para Login</a>
+        <a href="{{ url_for('auth.login') }}" class="btn btn-success"><span class="glyphicon glyphicon-log-in"></span> Ir para Login</a>
     </div>
 {% endblock content %}
